# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto APIs
  description: WASAPhoto app backend REST APIs
  version: '3'
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: |-
        The request was not compliant with the documentation
        (eg. missing fields, etc).
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  parameters:
    UserID:
      schema:
        type: integer
        example: 1
        readOnly: true
      name: userId
      in: path
      required: true
      description: User ID
    CommentId:
      schema:
        type: integer
        example: 1
        readOnly: true
      name: commentId
      in: path
      required: true
      description: Comment id
    FollowedId:
      schema:
        type: integer
        example: 1
        readOnly: true
      name: followedId
      in: path
      required: true
      description: User followed Id
...
    PhotoId:
      schema:
        type: integer
        example: 22
        readOnly: true
      name: photoId
      in: path
      required: true
      description: Photo id
  schemas:
...
    IdPhoto:
      description: Photo identifier
      type: integer
      minimum: 1
      maximum: 9999999
      example: 123456
    nComments:
      description: Represent a number of comments posted for one photo
      type: integer
      minimum: 0
      maximum: 9999999
      example: 123
    nLikes:
      description: Represent a number of likes added for one photo
      type: integer
      minimum: 0
      maximum: 9999999
      example: 123
...
security:
  - bearerAuth: []      
paths:
...
  /photos:
    post:
      operationId: uploadPhoto
      summary: Uploads a photo
      tags:
        - Photo
      description: Returns the identifier of a uploaded photo
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              description: Represents a photo posted by user
              properties:
                image:
                  type: string
                  description: Image to post
                  minLength: 0
                  maxLength: 999999
                  format: binary
                idUser:
                  description: The username that posts the photo
                  type: integer
                  example: 55
      responses:
        '201':
          description: Returns the id o a photo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdPhoto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/following/{followedId}:
    parameters:
      - $ref: '#/components/parameters/UserID'
      - $ref: '#/components/parameters/FollowedId'
    put:
      operationId: followUser
      summary: Follows a user
      tags:
        - Followers
      description: >-
        The user identified as "userId" follows the user identified as
        "followedId". Return the number of followed users by the user
      responses:
        '201':
          description: Returns the number of users followed by the user after the operation
          content:
            application/json:
              schema:
                description: Represent number of users followed
                type: integer
                example: 1234
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: unfollowUser
      summary: Unfollows a user
      description: Returns the number of followed users by the user after the operation
      tags:
        - Followers
      responses:
        '200':
          description: Returns the number of users followed by the user after the operation
          content:
            application/json:
              schema:
                description: Represent number of users followed
                type: integer
                example: 1234
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{photoId}/likes/{userId}:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
      - $ref: '#/components/parameters/UserID'
    put:
      operationId: likePhoto
      summary: Likes a photo of one user
      description: Likes a photo. Returns the number of likes of the photo after the "like"
      tags:
        - Like
      responses:
        '200':
          description: Number of likes of the post after the operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nLikes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: unlikePhoto
      summary: Unlikes a photo
      description: >-
        Unlikes a photo. Returns the number of likes of the photo after the
        dislike
      tags:
        - Like
      responses:
        '200':
          description: Number of likes of the post after the unlike
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nLikes'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /photos/{photoId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/PhotoId'
      - $ref: '#/components/parameters/CommentId'
    delete:
      summary: Remove the comment of one post
      description: >
        Remove the comment "commentId" of the photo "photoId"
      operationId: uncommentPhoto
      tags:
       - Comment
      responses:
        '200':
          description: Returns the number of comments after the operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/nComments'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...